WIP: Cleanup integration tests
