--- contrib/pg_upgrade/gp/Makefile
@@ -1,6 +1,0 @@
-# top_builddir = ../../..
-#
-# OBJS = check_greenplum.o checks.o checks_list.o checks_utilities.o
-#
-# include $(top_builddir)/src/Makefile.global
-# include $(top_srcdir)/contrib/contrib-global.mk
