+++ contrib/pg_upgrade/gp/test/check_covering_aoindex_test.c
@@ -1,90 +1,15 @@
+/*
+ * Integration tests for check_covering_aoindex()
+ */
+
 #include <stdarg.h>
 #include <stddef.h>
 #include <setjmp.h>
 
 #include "cmockery.h"
-#include "stdbool.h"
 
-
+#include "test_utils.h"
 #include "../checks.h"
-#include "../../pg_upgrade.h"
-#include "libpq-fe.h"
-
-ClusterInfo old_cluster;
-ClusterInfo new_cluster;
-OSInfo os_info;
-UserOpts user_opts;
-
-
-bool
-exec_prog(const char *log_file, const char *opt_log_file,
-          bool throw_error, const char *fmt,...){
-	return true;
-}
-
-
-/*
- * executeQueryOrDie()
- *
- *	Formats a query string from the given arguments and executes the
- *	resulting query.  If the query fails, this function logs an error
- *	message and calls exit() to kill the program.
- */
-static void
-executeQuery(PGconn *conn, const char *fmt,...)
-{
-	static char command[8192];
-	va_list		args;
-
-	va_start(args, fmt);
-	vsnprintf(command, sizeof(command), fmt, args);
-	va_end(args);
-
-	PQexec(conn, command);
-}
-
-static void 
-setup_old_cluster()
-{
-	char *port_string = getenv("PGPORTOLD");
-
-
-	if (!port_string)
-	{
-		printf("Must set PGPORTOLD (port number of 5X server) to run tests.\n");
-		exit(1);
-	}
-
-
-	old_cluster.port = atoi(port_string);
-	old_cluster.major_version = 80300;
-}
-
-
-static char * 
-get_database_name()
-{
-	return "postgres";
-}
-
-
-void setup_os_info()
-{
-	os_info.user = "";
-}
-
-static void 
-enable_utility_mode()
-{
-	old_cluster.use_utility_mode = true;
-}
-
-
-static void
-disable_utility_mode()
-{
-	old_cluster.use_utility_mode = false;
-}
 
 
 static void
